Learning Bistro
Free Open Source Website on which the student and teacher can go through from Overview to the depth of Python and Java.
The website would be useful for those who are self-learner or those who have the anxiety to learn.
Teacher can also go through the website to prepare the material for students and give a path to learn the languages.